Added modifyOutput
Added a StateMonad operator that will modify both the state and output based only on state.  Note that this command calls both executeCom and outputCom.

Fixed the StateMonad operator that will modify both the state and output based on previous output and state.  The most general of the state modification functions.